Rails angular tood
Re-arrange code to have everything in the root in the Angular app (it
is still /todos in the Rails API).  Still needs more work to integrate
the edit / add screen into the main screen and change the screen to
say "update" instead of "add".  Either change the internal route or
create a second one for edit.